Display app to users after it finishes loading
Previously a big white, empty window (with a smaller black rectangle) was displayed for a moment.
Now once the app is ready, a new `BrowserWindow` is created as hidden (`show: false`) and the index.html is loaded. After the index.html is loaded the window is shown to user and maximized.